Fixed RunInTransactionAsync to prevent deadlocks
Using SQLiteAsyncConnection inside RunInTransactionAsync's delegate
caused a deadlock when any task started inside its method calls ended up
in a different thread - both threads were trying to obtain a lock on the
same connection.
The new method passes SQLiteConnection to the delegate instead,
preventing this situation. The delegate doesn't really need to obtain a
lock since it runs synchronously with the rest of the transaction.
The delegates are also easier to write now without the need to call
Wait() at the end of each  asynchronous database call inside it. The
complete transaction still runs asynchronously to the calling code as it
should.